requirements.txt
Pillow
numpy
webdataset
torch
torchvision
timm
vllm
transformers
accelerate
huggingface_hub